The android code actually passes a full test! But Java won't compile (yet)
I have started on the Java code (which is completely broken) and got the
Android code to actually run a full hidden service end to end test!

Redid the android test structure so we could control when the
OnionProxyManager is created. This was important for making it easy to
test a 'fresh' install (by deleting the old file directory) before calling
the constructor.

Of the actual tests only testCleanInstall (and only in Android) is real.
I will fill in the rest of the tests later once I get things going in Java.

Changed the ports in torrc to make sure we don't collide with Briar

Got rid of the dedicated EventHandler class because it turns out that there
is state in OnionProxyManager (like bootstrapped) that is controlled by
data from the events. So I put the EventHandler callbacks back into
OnionProxyManager

Fixed up the Java build.gradle to copy over libs, tests, code & assets
from Android to Java so we can share them

Started in on the various Java classes but right now things won't even
compile.