Major file handling clean up
Changed the logic so that knowledge of where files go and how to install
them is centralized inside the OnionProxyContext object.

Unzipped the Android and Windows tor binaries since they are unitary.
But we are zipping the Mac binaries because there are two of them and
there is no non-icky way to enumerate the contents of a resource folder.

Brough back the unzipping function to handle Mac.