Preparing for Linux Support & rolling back Android
I changed the Android gradle and build tools back to their older versions
because the new Versions gave IntelliJ serious indigestion

I also made the environment variables for the Tor Process into a context
item because we need an additional environment variable on Linux.

I also threw in my first draft of the code to detect linux type. Next
stop will be to actually test it on Linux!