Refactored into three projects
I needed to create a 'universal' project that contains all the code shared
between Android and Java so that the other 'universal' projects in Thali
could inherit this without having to inherit any 'android' or 'java'
specific content.

Note that tests still live in Android and get copied over to Java.