Get us on the global Thali build infrastructure
Get project to use same build infrastructure as the rest of Thali, see
the build.gradle files. Also updated the README to include this data.

Upgrade local gradle to 1.12

Change tor context constructor to take file object, this presumes we
will figure the file location outside the constructor. This was useful
for debugging.